move tests to smr.rs
