+++ yew/src/virtual_dom/vtag.rs
@@ -554,8 +554,6 @@ mod tests {
     use super::*;
     use crate::html;
     use std::any::TypeId;
-    #[cfg(feature = "sans_mount_render")]
-    use std::convert::TryFrom;
     #[cfg(feature = "std_web")]
     use stdweb::web::{document, IElement};
     #[cfg(feature = "wasm_test")]
@@ -966,92 +964,6 @@ mod tests {
 
         html! { <div><a data-val=<u32 as Default>::default() /> </div> };
         html! { <div><a data-val=Box::<u32>::default() /></div> };
-    }
-
-    #[test]
-    #[cfg(feature = "sans_mount_render")]
-    fn it_stringifies_simple() {
-        let p = html! {
-            <p></p>
-        };
-
-        if let VNode::VTag(p) = p {
-            let p_html = Html::try_from(*p)
-                .expect("HTML stringify error")
-                .to_string();
-
-            assert_eq!(p_html, "<p />");
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "sans_mount_render")]
-    fn it_stringifies_complex() {
-        let other_sym = "bar";
-        let div = html! {
-            <div class=("foo", other_sym)>
-                { "quux" }
-            </div>
-        };
-        let p = html! {
-            <p aria-controls="it-works">
-                { "test" }
-                {div}
-            </p>
-        };
-
-        if let VNode::VTag(p) = p {
-            let p_html = Html::try_from(*p)
-                .expect("HTML stringify error")
-                .to_string();
-
-            assert_eq!(
-                p_html,
-                "<p aria-controls=\"it&#x2D;works\">test<div class=\"foo&#x20;bar\">quux</div></p>"
-            );
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "sans_mount_render")]
-    fn it_stringifies_attrs() {
-        let div = html! {
-            <div a="b" b="a" />
-        };
-
-        if let VNode::VTag(div) = div {
-            let div_html = Html::try_from(*div)
-                .expect("HTML stringify error")
-                .to_string();
-            let order_1 = "<div a=\"b\" b=\"a\" />";
-            let order_2 = "<div b=\"a\" a=\"b\" />";
-            assert!(div_html == order_1 || div_html == order_2);
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "sans_mount_render")]
-    fn it_does_not_stringify_special_attrs() {
-        let node_ref = NodeRef::default();
-
-        let div = html! {
-            <div ref=node_ref />
-        };
-
-        if let VNode::VTag(div) = div {
-            let div_html = Html::try_from(*div)
-                .expect("HTML stringify error")
-                .to_string();
-            assert_eq!(div_html, "<div />");
-        } else {
-            assert!(false);
-        }
     }
 
     #[test]
